+++ bundle-es2015/client/client.js
@@ -5,16 +5,22 @@ import { ClientConfig } from './client-config';
 import { NodeOperations } from '../nodes/node-operations';
 import { HttpClient } from 'contensis-core-api';
 import { ProjectOperations } from '../projects/project-operations';
-const Scopes = 'Entry_Read Entry_Write Entry_Delete ContentType_Read Project_Read';
+import { RoleOperations } from '../roles/RolesOperations';
+import { PermissionOperations } from '../permissions/permission-operations';
+import { ComponentOperations } from '../components/component-operations';
+const Scopes = 'ContentType_Read ContentType_Write ContentType_Delete Entry_Read Entry_Write Entry_Delete Project_Read Project_Write Project_Delete';
 export class Client {
     constructor(config = null) {
         this.clientConfig = null;
         this.clientConfig = new ClientConfig(config, Client.defaultClientConfig);
         this.httpClient = new HttpClient(this);
+        this.components = new ComponentOperations(this.httpClient, this);
         this.contentTypes = new ContentTypeOperations(this.httpClient, this);
         this.entries = new EntryOperations(this.httpClient, this);
         this.nodes = new NodeOperations(this.httpClient, this);
+        this.permissions = new PermissionOperations(this.httpClient, this);
         this.projects = new ProjectOperations(this.httpClient, this);
+        this.roles = new RoleOperations(this.httpClient, this);
     }
     static create(config = null) {
         return new Client(config);