+++ bundle-es2015/models/ComponentGetOptions.d.ts
@@ -1,0 +1,3 @@
+import { ContentTypeGetOptions } from './ContentTypeGetOptions';
+export interface ComponentGetOptions extends ContentTypeGetOptions {
+}