+++ bundle-es2015/models/Permission.d.ts
@@ -1,0 +1,5 @@
+export interface Permission {
+    id: string;
+    languages: string[];
+    actions: string[];
+}