+++ bundle-es2015/models/PermissionGetOptions.d.ts
@@ -1,0 +1,7 @@
+import { ResourceType } from './ResourceType';
+export interface PermissionGetOptions {
+    resourceType: ResourceType;
+    resourceId: string;
+    userId?: string;
+    language?: string;
+}