+++ bundle-es2015/models/Role.d.ts
@@ -1,0 +1,14 @@
+import { Permissions } from './Permissions';
+import { Assignments } from './Assignments';
+export interface Role {
+    id: string;
+    name: {
+        [key: string]: string;
+    };
+    description?: {
+        [key: string]: string;
+    };
+    enabled: boolean;
+    permissions: Permissions;
+    assignments: Assignments;
+}