+++ lib/models/Assignments.d.ts
@@ -1,0 +1,5 @@
+export interface Assignments {
+    users?: string[];
+    groups?: string[];
+    apiKeys?: string[];
+}