+++ src/client/client.ts
@@ -1,7 +1,7 @@
 import fetch from 'cross-fetch';
 import {
 	Config, ContensisClient, IContentTypeOperations,
-	IEntryOperations, INodeOperations, IProjectOperations
+	IEntryOperations, INodeOperations, IProjectOperations, IRoleOperations, IPermissionOperations, IComponentOperations
 } from '../models';
 import { EntryOperations } from '../entries/entry-operations';
 import { ContentTypeOperations } from '../content-types/content-type-operations';
@@ -9,18 +9,24 @@ import { ClientConfig } from './client-config';
 import { NodeOperations } from '../nodes/node-operations';
 import { ClientParams, HttpClient, IHttpClient } from 'contensis-core-api';
 import { ProjectOperations } from '../projects/project-operations';
+import { RoleOperations } from '../roles/RolesOperations';
+import { PermissionOperations } from '../permissions/permission-operations';
+import { ComponentOperations } from '../components/component-operations';
 
-const Scopes = 'Entry_Read Entry_Write Entry_Delete ContentType_Read Project_Read';
+const Scopes = 'ContentType_Read ContentType_Write ContentType_Delete Entry_Read Entry_Write Entry_Delete Project_Read Project_Write Project_Delete';
 
 export class Client implements ContensisClient {
 
 	static defaultClientConfig: ClientConfig = null;
 
 	clientConfig: ClientConfig = null;
-	entries: IEntryOperations;
+	components: IComponentOperations;
 	contentTypes: IContentTypeOperations;
+	entries: IEntryOperations;
 	nodes: INodeOperations;
+	permissions: IPermissionOperations;
 	projects: IProjectOperations;
+	roles: IRoleOperations;
 
 	private httpClient: IHttpClient;
 	private token: string;
@@ -37,10 +43,13 @@ export class Client implements ContensisClient {
 		this.clientConfig = new ClientConfig(config, Client.defaultClientConfig);
 		this.httpClient = new HttpClient(this);
 
+		this.components = new ComponentOperations(this.httpClient, this);
 		this.contentTypes = new ContentTypeOperations(this.httpClient, this);
 		this.entries = new EntryOperations(this.httpClient, this);
 		this.nodes = new NodeOperations(this.httpClient, this);
+		this.permissions = new PermissionOperations(this.httpClient, this);
 		this.projects = new ProjectOperations(this.httpClient, this);
+		this.roles = new RoleOperations(this.httpClient, this);
 	}
 
 	public getParams(): ClientParams {