+++ src/models/AuthorizationGetOptions.ts
@@ -1,0 +1,5 @@
+import { PermissionGetOptions } from './PermissionGetOptions';
+
+export interface AuthorizationGetOptions extends PermissionGetOptions {
+    actionName: string;
+}