+++ src/models/ComponentGetOptions.ts
@@ -1,0 +1,4 @@
+import { ContentTypeGetOptions } from './ContentTypeGetOptions';
+
+export interface ComponentGetOptions extends ContentTypeGetOptions {
+}