+++ src/models/PermissionGetOptions.ts
@@ -1,0 +1,8 @@
+import { ResourceType } from './ResourceType';
+
+export interface PermissionGetOptions {
+    resourceType: ResourceType;
+    resourceId: string;
+    userId?: string;
+    language?: string;
+}