+++ src/models/Permissions.ts
@@ -1,0 +1,6 @@
+import { Permission } from './Permission';
+
+export interface Permissions {
+    contentTypes?: Permission[];
+    entries?: Permission[];
+}