+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-beta.4] - 2019-12-11
+### Added
+- full api support for components, roles and permissions;
+
 ## [1.0.0-beta.3] - 2019-12-06
 ### Added
-- complete api support for projects, content types, entries and nodes;
+- full api support for projects, content types, entries and nodes;
 
 ## [1.0.0-beta.2] - 2019-11-28
 ### Added